Edit files for A-Packages component (organize files into packages)
