+++ src/main/java/mike/Main.java
@@ -1,3 +1,5 @@
+package mike;
+
 public class Main {
 
     public static void main(String[] args) {
