+++ src/main/java/mike/Mike.java
@@ -1,3 +1,5 @@
+package mike;
+
 import java.time.format.DateTimeParseException;
 import java.util.Scanner;
