+++ src/main/java/mike/Storage.java
@@ -1,3 +1,5 @@
+package mike;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
