+++ src/main/java/mike/Ui.java
@@ -1,3 +1,5 @@
+package mike;
+
 public class Ui {
 
     public Ui () {
